requirements.txt
# python -m spacy download es_core_news_md/en_core_web_md
# pip install setuptools --upgrade --ignore-installed

binaryornot==0.4.4
boto3 #no-deploy
botocore #no-deploy
absl-py==0.9.0
alembic==1.3.2
aniso8601==8.0.0
astor==0.8.1
asn1crypto==1.3.0
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
cytoolz==0.10.1
decorator==4.4.1
dill==0.3.1.1
elasticsearch==7.1.0
flasgger==0.9.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
google-pasta==0.1.8
goslate==1.5.1
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
healthcheck==1.3.3
hypothesis==5.1.5
idna==2.8
importlib-metadata==1.4.0
imutils==0.5.3
itsdangerous==1.1.0
jaeger-client==4.2.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
#Keras==2.3.1
#Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==8.1.0
msgpack==0.6.2
msgpack-numpy==0.4.4.3
murmurhash==1.0.2
numpy==1.18.1
opencv-python==4.1.2.30
opentracing==2.3.0
opentracing-utils==0.18.1
packaging==20.0
pandas==0.25.3
parso==0.5.2
pdf2image==1.11.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prompt-toolkit==3.0.2
protobuf==3.11.2
py==1.8.1
pycparser==2.19
PyDictionary==1.5.2
Pygments==2.5.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyreadline==2.1
PySocks==1.7.1
pytest==5.3.2
python-dateutil==2.8.1
python-editor==1.0.4
python-Levenshtein==0.12.0
pytz==2019.3
PyYAML==5.3
recordtype==1.3
regex==2020.1.8
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
soupsieve==1.9.5
spacy==2.2.3
SQLAlchemy==1.3.12
tensorflow==2.0.0
termcolor==1.1.0
threadloop==1.0.2
thrift==0.13.0
tika==1.23
toolz==0.10.0
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
ujson==1.35
urllib3==1.25.7
urllib5==5.0.0
urlparse2==1.1.1
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
wincertstore==0.2
win-inet-pton==1.1.0
wrapt==1.11.2
zipp==0.6.0
